requirements.txt
# Python 3.12.7 compatible dependencies

# Core
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.1
pydantic>=2.9.0

# LangChain Ecosystem (Pinecone & AWS)
langchain>=0.3.0
langchain-core>=0.3.10
langchain-community>=0.3.0
langchain-aws>=0.1.1

# Pinecone
#pinecone-client>=5.0.0
pinecone>=5.0.0

# NLP / Embeddings
tiktoken>=0.7.0
numpy>=1.26.0
pandas>=2.2.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
requests>=2.32.3

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.2
html2text>=2024.2.26

# Web UI
streamlit>=1.39.0

fpdf2>=2.8.1
pypdf>=4.2.0


